+++ src/main/java/com/prgrms/catchtable/shop/dto/response/GetShopResponse.java
@@ -1,4 +1,4 @@
-package com.prgrms.catchtable.shop.dto;
+package com.prgrms.catchtable.shop.dto.response;
 
 import java.math.BigDecimal;
 import java.time.LocalTime;
