+++ src/main/java/com/prgrms/catchtable/shop/dto/response/RegistShopResponse.java
@@ -1,4 +1,4 @@
-package com.prgrms.catchtable.shop.dto;
+package com.prgrms.catchtable.shop.dto.response;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
 import java.time.LocalTime;
